[IMP] 优化查询maven settins逻辑

See merge request hand-yanqianyun-yq-devops/devops-service!565

+++ src/main/java/io/choerodon/devops/app/service/impl/DevopsHostServiceImpl.java
@@ -5,15 +5,20 @@ import java.io.InputStream;
 import java.nio.charset.StandardCharsets;
 import java.util.*;
 import java.util.function.Function;
+import java.util.function.Predicate;
+import java.util.function.Supplier;
 import java.util.stream.Collectors;
 import javax.annotation.Nullable;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.google.gson.Gson;
+import io.choerodon.devops.infra.dto.workflow.DevopsPipelineDTO;
 import net.schmizz.sshj.SSHClient;
 import net.schmizz.sshj.common.IOUtils;
 import org.apache.commons.lang3.StringUtils;
+import org.hzero.core.util.AssertUtils;
 import org.hzero.websocket.helper.KeySocketSendHelper;
+import org.jetbrains.annotations.NotNull;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Lazy;
@@ -219,9 +224,14 @@ public class DevopsHostServiceImpl implements DevopsHostService {
     @Override
     public void deleteHost(Long projectId, Long hostId) {
         DevopsHostDTO devopsHostDTO = devopsHostMapper.selectByPrimaryKey(hostId);
+        if (devopsHostDTO == null) return;
         devopsHostUserPermissionService.checkUserOwnManagePermissionOrThrow(projectId, devopsHostDTO, DetailsHelper.getUserDetails().getUserId());
-        checkEnableHostDelete(hostId);
-        CommonExAssertUtil.assertTrue(devopsHostDTO.getProjectId().equals(projectId), MiscConstants.DEVOPS_OPERATING_RESOURCE_IN_OTHER_PROJECT);
+        //不在校验主机的连接状态
+//        checkEnableHostDelete(hostId);
+        //校验流水线是否引用了该主机
+        List<CiCdPipelineDTO> ciCdPipelineDTOS = devopsHostMapper.selectPipelineByHostId(hostId);
+        AssertUtils.isTrue(CollectionUtils.isEmpty(ciCdPipelineDTOS), handHostCheckMsg(ciCdPipelineDTOS));
+        CommonExAssertUtil.assertTrue(devopsHostDTO.getProjectId().equals(projectId), MiscConstants.ERROR_OPERATING_RESOURCE_IN_OTHER_PROJECT);
         try {
             devopsHostMapper.deleteByPrimaryKey(hostId);
 //            devopsDockerInstanceMapper.deleteByHostId(hostId);
@@ -368,13 +378,10 @@ public class DevopsHostServiceImpl implements DevopsHostService {
     @Override
     public boolean checkHostDelete(Long projectId, Long hostId) {
         DevopsHostDTO devopsHostDTO = devopsHostMapper.selectByPrimaryKey(hostId);
-        if (Objects.isNull(devopsHostDTO)) {
-            throw new CommonException(ERROR_HOST_NOT_FOUND);
-        }
-        if (DevopsHostStatus.CONNECTED.getValue().equals(devopsHostDTO.getHostStatus())) {
-            return Boolean.FALSE;
-        }
-
+        if (Objects.isNull(devopsHostDTO)) return Boolean.TRUE;
+        //主机关联流水线任务不能删除
+        List<CiCdPipelineDTO> ciCdPipelineDTOS = devopsHostMapper.selectPipelineByHostId(hostId);
+        if (!CollectionUtils.isEmpty(ciCdPipelineDTOS)) return Boolean.FALSE;
         return Boolean.TRUE;
     }
 
@@ -888,5 +895,12 @@ public class DevopsHostServiceImpl implements DevopsHostService {
         map.put("status", status);
         map.put("exception", exception);
         return map;
+    }
+
+    @NotNull
+    private static Supplier<String> handHostCheckMsg(List<CiCdPipelineDTO> ciCdPipelineDTOS) {
+        return () -> {
+            throw new CommonException("error.host.linked.pipeline.delete", ciCdPipelineDTOS.get(0) == null ? "" : ciCdPipelineDTOS.get(0).getName());
+        };
     }
 }
