+++ src/main/resources/shell/ci.sh
@@ -357,12 +357,44 @@ function saveJarMetadata() {
     -o "${CI_COMMIT_SHA}-ci.response" \
     -w %{http_code})
   # 判断本次上传到devops是否出错
-  response_upload_to_devops=$(cat "${CI_COMMIT_SHA}-ci.response")
-  rm "${CI_COMMIT_SHA}-ci.response"
-  if [ "$result_upload_to_devops" != "200" ]; then
-    echo "$response_upload_to_devops"
-    echo "upload to devops error"
-    exit 1
+  if [ -e "${CI_COMMIT_SHA}-ci.response" ]; then
+    response_upload_to_devops=$(cat "${CI_COMMIT_SHA}-ci.response")
+    rm "${CI_COMMIT_SHA}-ci.response"
+    if [ "$result_upload_to_devops" != "200" ]; then
+      echo $response_upload_to_devops
+      echo "upload to devops error"
+      exit 1
+    fi
+  fi
+}
+
+############################### 存储jar包元数据, 用于CD阶段主机部署-jar包部署 ################################
+# $1 maven制品库id
+# $2 mvn_settings_id    mvn_settings_id
+# $3 sequence   猪齿鱼的CI流水线的步骤的序列号
+function saveJarInfo() {
+  result_upload_to_devops=$(curl -X POST \
+    -H 'Expect:' \
+    -F "token=${Token}" \
+    -F "nexus_repo_id=$1" \
+    -F "mvn_settings_id=$2" \
+    -F "sequence=$3" \
+    -F "gitlab_pipeline_id=${CI_PIPELINE_ID}" \
+    -F "job_name=${CI_JOB_NAME}" \
+    -F "version=${CI_COMMIT_TAG}" \
+    -F "file=@pom.xml" \
+    "${CHOERODON_URL}/devops/ci/save_jar_info" \
+    -o "${CI_COMMIT_SHA}-ci.response" \
+    -w %{http_code})
+  # 判断本次上传到devops是否出错
+  if [ -e "${CI_COMMIT_SHA}-ci.response" ]; then
+    response_upload_to_devops=$(cat "${CI_COMMIT_SHA}-ci.response")
+    rm "${CI_COMMIT_SHA}-ci.response"
+    if [ "$result_upload_to_devops" != "200" ]; then
+      echo $response_upload_to_devops
+      echo "upload to devops error"
+      exit 1
+    fi
   fi
 }
 
@@ -415,12 +447,47 @@ function saveCustomJarMetadata() {
     -o "${CI_COMMIT_SHA}-ci.response" \
     -w %{http_code})
   # 判断本次上传到devops是否出错
-  response_upload_to_devops=$(cat "${CI_COMMIT_SHA}-ci.response")
-  rm "${CI_COMMIT_SHA}-ci.response"
-  if [ "$result_upload_to_devops" != "200" ]; then
-    echo "$response_upload_to_devops"
-    echo "upload to devops error"
-    exit 1
+  if [ -e "${CI_COMMIT_SHA}-ci.response" ]; then
+    response_upload_to_devops=$(cat "${CI_COMMIT_SHA}-ci.response")
+    rm "${CI_COMMIT_SHA}-ci.response"
+    if [ "$result_upload_to_devops" != "200" ]; then
+      echo $response_upload_to_devops
+      echo "upload to devops error"
+      exit 1
+    fi
+  fi
+}
+############################### 存储jar包元数据, 用于CD阶段主机部署-jar包部署 ################################
+# $1 mvn_settings_id   mvn_settings_id
+# $2 sequence   猪齿鱼的CI流水线的步骤的序列号
+# $3 maven_repo_url   目标仓库地址
+# $4 username   目标仓库用户名
+# $5 password   目标仓库用户密码
+function saveCustomJarInfo() {
+  result_upload_to_devops=$(curl -X POST \
+    -H 'Expect:' \
+    -F "token=${Token}" \
+    -F "mvn_settings_id=$1" \
+    -F "sequence=$2" \
+    -F "maven_repo_url=$3" \
+    -F "username='$4'" \
+    -F "password='$5'" \
+    -F "gitlab_pipeline_id=${CI_PIPELINE_ID}" \
+    -F "job_name=${CI_JOB_NAME}" \
+    -F "version=${CI_COMMIT_TAG}" \
+    -F "file=@pom.xml" \
+    "${CHOERODON_URL}/devops/ci/save_jar_info" \
+    -o "${CI_COMMIT_SHA}-ci.response" \
+    -w %{http_code})
+  # 判断本次上传到devops是否出错
+  if [ -e "${CI_COMMIT_SHA}-ci.response" ]; then
+    response_upload_to_devops=$(cat "${CI_COMMIT_SHA}-ci.response")
+    rm "${CI_COMMIT_SHA}-ci.response"
+    if [ "$result_upload_to_devops" != "200" ]; then
+      echo $response_upload_to_devops
+      echo "upload to devops error"
+      exit 1
+    fi
   fi
 }
 ############################### 存储jar包元数据, 用于CD阶段主机部署-jar包部署 ################################
@@ -618,6 +685,8 @@ function rewrite_image_info() {
       export HARBOR_CONFIG_ID=$(jq -r .harborRepoId rewrite_image_info.json)
       export REPO_TYPE=$(jq -r .repoType rewrite_image_info.json)
       export DOCKER_PASSWORD=$(jq -r .dockerPassword rewrite_image_info.json)
+      # 设成docekr认证配置文件
+      echo "{\"auths\":{\"$DOCKER_REGISTRY\":{\"auth\":\"$(echo -n $DOCKER_USERNAME:$DOCKER_PASSWORD | base64)\"}}}" | tr -d '\n' > $DOCKER_CONFIG/config.json
     fi
   fi
